+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+### 2021-05-21
+
+* Fix localization functionality according to the geopose protocol
+* Fix some bugs (rotationDevice, more debug info)
 
 ### 2021-04-29